Removed variable rate case and some fixes.
I remove the event separation with a variable rate here. Events were separated according to the given rate distribution but the absolute rates were only matched if the number of simulated events magically matches the number of events given in the rate evolution csv file. We can implement this function later with proper normalization.